Added a sample of the animals controller
This sample is from the main controller class for the animals movements.
It does not have all methods and attributes but the original class contains all methods and attributes to choose random walking paths for animals, perform obstacle avoidance and make the animals walk or idle. I refactored it. 
It also contains XML code for documentation.